Fix type mismatches in test assertions for standard testify
Standard testify does strict type comparison in assert.Equal.
Cast expected values to match actual types:
- float32 for GetWatermark results
- os.FileMode for file mode bitwise ops
- float64 for NsPerOpDelta

+++ src/bench/compare_test.go
@@ -28,7 +28,7 @@ func TestCompareNoPrevious(t *testing.T) {
 	require.Equal(t, 1, len(deltas))
 
 	assert.Nil(t, deltas[0].Previous)
-	assert.Equal(t, 0, deltas[0].NsPerOpDelta)
+	assert.Equal(t, float64(0), deltas[0].NsPerOpDelta)
 }
 
 func TestCompareWithPrevious(t *testing.T) {
